+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Use positional parameters when applicable
 * Merge evaluator class with inc class
 * Use native shell execution instead of Python
+* Trim output flag
 
 ## 1.0.1+1